Extract accordion-highlight state to shared module (#1346 step 2f)
This is the first PR of step 2f in issue #1346 -- breaking the
eventHandler.js <-> overlayHandler.js circular reference that has
been blocking the rest of the ES-modules migration.

## Why this unblocks Step 9

The circular ref today: eventHandler.js exposes
window.EventHandler.updateAccordionHighlights, which is called from
overlayHandler.js (5 sites) and 025-libraries.js (5 sites).
Meanwhile, overlayHandler exposes window.OverlayHandler methods
that eventHandler needs to call. Neither file can convert to a
clean `import` topology because they'd import each other.

Extracting the accordion-highlight state to a third module resolves
the cycle: overlayHandler.js and 025-libraries.js can each import
accordionHighlights independently, without eventHandler needing to
be a dependency at all. That's the pattern Step 9 will need to
convert overlayHandler.js and 025-libraries.js into Svelte
component islands (they need clean import boundaries first).

## What this PR does

1. Creates `static/local-js/modules/accordionHighlights.js` (295
   lines) exporting 5 pure functions that own `.selected` class
   state on `.accordion-header` elements:

     hasCheckedTemplateGroupToggle(body)  -- predicate, tri-state
     hasLibraryFileEntries(body)          -- predicate, boolean
     highlightParentAccordions(el)        -- ancestor walker
     removeHighlightIfEmpty(el)           -- ancestor cleanup
     updateAccordionHighlights()          -- top-level sweep

2. Removes those 5 methods from `EventHandler` in eventHandler.js
   (-233 lines).

3. Re-attaches them onto the `EventHandler` object at import time:

     import * as accordionHighlights from './modules/accordionHighlights.js'
     ...
     EventHandler.updateAccordionHighlights = accordionHighlights.updateAccordionHighlights
     // (etc)
     window.EventHandler = EventHandler

   Non-module consumers that reach for
   `window.EventHandler.updateAccordionHighlights()` (overlayHandler,
   025-libraries) keep working unchanged. Follow-up PRs will replace
   those callsites with direct imports.

## Verbatim behavior preservation

- Every function body was moved character-for-character.
- Console-log emojis preserved via `\u{XXXX}` escapes.
- Inter-function references rewritten from `EventHandler.foo(...)`
  to bare `foo(...)` (they're now in the same module scope).
- removeHighlightIfEmpty had a local `const hasLibraryFileEntries`
  that shadowed the module function name; renamed the local to
  `bodyHasLibraryFileEntries` to avoid the shadow (same behavior,
  clearer read).
- The self-recursive call at the end of removeHighlightIfEmpty
  still recurses on itself (bare call, same module).

## eventHandler.js: 768 -> 548 lines (-220 net)

The removed EventHandler methods add 5 lines of monkey-patch and
1 line of import at the top, so the net-net for eventHandler.js is
-220 lines. Compare to the module: 295 lines including
docstrings that were spread thin across the original file.

## Test coverage

- Existing tests/js/eventHandler.test.js (40 tests) still passes.
  Every test that called `EventHandler.updateAccordionHighlights()`
  or the other 4 methods is now silently going through the module.
  That's the strongest possible verbatim-behavior check.

- New tests/js/modules/accordionHighlights.test.js (14 tests):
  * Module surface: 6 tests. Prove all 5 functions are exported and
    nothing else (namespace-keys assertion catches accidental leaks).
  * Standalone smoke tests: 8 tests. One per function on the
    minimum DOM fixture. Prove the module works without
    eventHandler.js being imported.

## Verification

- 1498/1498 Vitest tests pass (was 1484; +14 new module tests)
- 5/5 relevant e2e tests pass (console-error smoke + library page)
- ESLint clean on all touched files

## Not in this PR (deliberately)

- overlayHandler.js and 025-libraries.js still reach for
  `window.EventHandler.updateAccordionHighlights`. Migrating those
  callsites to direct imports is a follow-up PR (mechanical,
  10 callsites total) and keeps THIS PR's diff focused on the
  extract. The `window.EventHandler` shim in eventHandler.js
  keeps them working in the meantime.

- The other half of the circular ref
  (`OverlayHandler.initializeOverlays` and
  `OverlayHandler.updateHiddenInputs` called from eventHandler.js)
  is a separate, smaller extract that will follow.
